\name{dFCHMM-package}
\alias{dFCHMM-package}
\alias{dFCHMM}
\docType{package}
\title{
\packageTitle{dFCHMM}
}
\description{
\packageDescription{dFCHMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dFCHMM}
\packageIndices{dFCHMM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dFCHMM}

Maintainer: \packageMaintainer{dFCHMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
